\name{delphe-package}
\alias{delphe-package}
\alias{delphe}
\docType{package}
\title{
DatabasE Linking for Population, Health and Environmental Scientists

}
\description{
Connecting to PostGIS database, sending queries, downloading data... other stuff.

}
\details{
\tabular{ll}{
Package: \tab delphe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-29\cr
License: \tab GPL2\cr
}

}
\author{
ivanhanigan

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{

}

\keyword{ package }
\seealso{

}
\examples{# connect2postgres()
}
